requirements.txt
streamlit==1.36.0
scikit-learn>=1.5.2
numpy>=1.26.4
pandas==2.2.2
python-dotenv==1.0.0
pinecone-client==4.0.0
chromadb==0.5.3
weaviate-client==4.6.0
elasticsearch==8.13.0
openai==1.37.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
tqdm==4.66.1
